Add a script for stage2 build
self.sh modifies source files for punyc so that they are
more friendly for self-hosting. The script is used to create
a punyc binary that consists of object files partly genearted
by gcc and the others by punyc.

punyc can now compile main.c and type.c with the script.